changed logic of listing nft on sale
