+++ market-contract/src/external.rs
@@ -2,10 +2,10 @@ use crate::*;
 
 /// external contract calls
 
-//initiate a cross contract call to the nft contract. This will transfer the token to the buyer and return
-//a payout object used for the market to distribute funds to the appropriate accounts.
+//initiate a cross contract call to the nft contract
 #[ext_contract(ext_contract)]
 trait ExtContract {
+    //This will transfer the token to the buyer and return a payout object used for the market to distribute funds to the appropriate accounts
     fn nft_transfer_payout(
         &mut self,
         receiver_id: AccountId, //purchaser (person to transfer the NFT to)
@@ -18,6 +18,13 @@ trait ExtContract {
         */
         balance: U128,
         //the maximum amount of accounts the market can payout at once (this is limited by GAS)
-		max_len_payout: u32,
+		    max_len_payout: u32,
+    );
+    fn nft_token(&self, token_id: TokenId);
+    fn nft_is_approved(
+        &self,
+        token_id: TokenId,
+        approved_account_id: AccountId,
+        approval_id: u64,
     );
 }
