+++ market-contract/src/sale.rs
@@ -1,5 +1,6 @@
 use crate::*;
-use near_sdk::promise_result_as_success;
+use near_sdk::{log, promise_result_as_success, PromiseError};
+use near_sdk::serde_json::json;
 
 //struct that holds important information about each sale on the market
 #[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize)]
@@ -17,9 +18,45 @@ pub struct Sale {
     pub sale_conditions: SalePriceInYoctoNear,
 }
 
+//The Json token is what will be returned from view calls. 
+#[derive(Serialize, Deserialize)]
+#[serde(crate = "near_sdk::serde")]
+pub struct JsonToken {
+    //owner of the token
+    pub owner_id: AccountId,
+}
+
+//struct for keeping track of the sale conditions for a Sale
+#[derive(Serialize, Deserialize)]
+#[serde(crate = "near_sdk::serde")]
+pub struct SaleArgs {
+    pub sale_conditions: SalePriceInYoctoNear,
+}
+
 #[near_bindgen]
 impl Contract {
-    
+    // lists a nft for sale on the market
+    #[payable]
+    pub fn list_nft_for_sale(
+      &mut self,
+      nft_contract_id: AccountId,
+      token_id: TokenId,
+      approval_id: u64,
+      msg: String,
+    ) {
+        let owner_id = env::predecessor_account_id();
+
+        let nft_token_promise = Promise::new(nft_contract_id.clone()).function_call(
+            "nft_token".to_owned(), json!({ "token_id": token_id }).to_string().into_bytes(), 0, Gas(10u64.pow(13))
+        );
+        let nft_is_approved_promise = Promise::new(nft_contract_id.clone()).function_call(
+          "nft_is_approved".to_owned(), json!({ "token_id": token_id, "approved_account_id": env::current_account_id(), "approval_id": approval_id }).to_string().into_bytes(), 0, Gas(10u64.pow(13))
+        );
+        nft_token_promise
+          .and(nft_is_approved_promise)
+          .then(Self::ext(env::current_account_id()).process_listing(owner_id.clone(), nft_contract_id, token_id, approval_id, msg));
+    }
+
     //removes a sale from the market. 
     #[payable]
     pub fn remove_sale(&mut self, nft_contract_id: AccountId, token_id: String) {
@@ -207,6 +244,114 @@ impl Contract {
 
         //return the price payout out
         price
+    }
+
+    #[private]
+    pub fn process_listing(
+        &mut self,
+        owner_id: AccountId,
+        nft_contract_id: AccountId,
+        token_id: TokenId,
+        approval_id: u64,
+        msg: String,
+        #[callback_result] nft_token_result: Result<JsonToken, PromiseError>,
+        #[callback_result] nft_is_approved_result: Result<bool, PromiseError>,
+    ) {
+        if let Ok(result) = nft_token_result {
+            assert_eq!(
+                result.owner_id,
+                owner_id,
+                "Signer is not NFT owner",
+            )
+        } else {
+            log!("nft_is_approved call failed");
+        }
+        if let Ok(result) = nft_is_approved_result {
+            assert_eq!(
+                result,
+                true,
+                "Marketplace contract is not approved",
+            )
+        } else {
+            log!("nft_is_approved call failed");
+        } 
+    
+        //we need to enforce that the user has enough storage for 1 EXTRA sale.
+
+        //get the storage for a sale. dot 0 converts from U128 to u128
+        let storage_amount = self.storage_minimum_balance().0;
+        //get the total storage paid by the owner
+        let owner_paid_storage = self.storage_deposits.get(&owner_id).unwrap_or(0);
+        //get the storage required which is simply the storage for the number of sales they have + 1 
+        let signer_storage_required = (self.get_supply_by_owner_id(owner_id.clone()).0 + 1) as u128 * storage_amount;
+        
+        //make sure that the total paid is >= the required storage
+        assert!(
+            owner_paid_storage >= signer_storage_required,
+            "Insufficient storage paid: {}, for {} sales at {} rate of per sale",
+            owner_paid_storage, signer_storage_required / STORAGE_PER_SALE, STORAGE_PER_SALE
+        );
+
+        //if all these checks pass we can create the sale conditions object.
+        let SaleArgs { sale_conditions } =
+            //the sale conditions come from the msg field. The market assumes that the user passed
+            //in a proper msg. If they didn't, it panics. 
+            near_sdk::serde_json::from_str(&msg).expect("Not valid SaleArgs");
+
+        //create the unique sale ID which is the contract + DELIMITER + token ID
+        let contract_and_token_id = format!("{}{}{}", nft_contract_id, DELIMETER, token_id);
+        
+        //insert the key value pair into the sales map. Key is the unique ID. value is the sale object
+        self.sales.insert(
+            &contract_and_token_id,
+            &Sale {
+                owner_id: owner_id.clone(), //owner of the sale / token
+                approval_id, //approval ID for that token that was given to the market
+                nft_contract_id: nft_contract_id.to_string(), //NFT contract the token was minted on
+                token_id: token_id.clone(), //the actual token ID
+                sale_conditions, //the sale conditions 
+          },
+        );
+
+        //Extra functionality that populates collections necessary for the view calls 
+
+        //get the sales by owner ID for the given owner. If there are none, we create a new empty set
+        let mut by_owner_id = self.by_owner_id.get(&owner_id).unwrap_or_else(|| {
+            UnorderedSet::new(
+                StorageKey::ByOwnerIdInner {
+                    //we get a new unique prefix for the collection by hashing the owner
+                    account_id_hash: hash_account_id(&owner_id),
+                }
+                .try_to_vec()
+                .unwrap(),
+            )
+        });
+        
+        //insert the unique sale ID into the set
+        by_owner_id.insert(&contract_and_token_id);
+        //insert that set back into the collection for the owner
+        self.by_owner_id.insert(&owner_id, &by_owner_id);
+
+        //get the token IDs for the given nft contract ID. If there are none, we create a new empty set
+        let mut by_nft_contract_id = self
+            .by_nft_contract_id
+            .get(&nft_contract_id)
+            .unwrap_or_else(|| {
+                UnorderedSet::new(
+                    StorageKey::ByNFTContractIdInner {
+                        //we get a new unique prefix for the collection by hashing the owner
+                        account_id_hash: hash_account_id(&nft_contract_id),
+                    }
+                    .try_to_vec()
+                    .unwrap(),
+                )
+            });
+        
+        //insert the token ID into the set
+        by_nft_contract_id.insert(&token_id);
+        //insert the set back into the collection for the given nft contract ID
+        self.by_nft_contract_id
+            .insert(&nft_contract_id, &by_nft_contract_id);
     }
 }
