--- rust-toolchain
@@ -1,1 +1,0 @@
-1.61.0
